requirements.txt
torch
transformers
datasets
peft
accelerate
bitsandbytes
trl
wandb  # optional logging
requests
python-dotenv

